# Conflicts:
#	coverage/.last_run.json
#	coverage/.resultset.json
#	coverage/index.html
#	spec/controllers/api/v1/brokers_controller_spec.rb

+++ app/controllers/api/v1/brokers_controller.rb
@@ -1,7 +1,8 @@
 module Api
   module V1
     class BrokersController < ApiController
-      before_action :check_current_company_requests , only: [:accept, :reject, :remove]
+      before_action :check_record_exit?, only: [:send_request]
+      before_action :check_current_company_requests, only: [:accept, :reject]
       skip_before_action :verify_authenticity_token, only: [:send_request, :dashboard, :company_record_on_the_basis_of_roles, :accept, :reject, :remove]
 
       include ApplicationHelper
@@ -11,12 +12,12 @@ module Api
         @data = []
         @sellers.each do |s|
           @data << {
-            id: s.id,
-            name: s.name
+              id: s.id,
+              name: s.name
           }
         end
         @all_sellers = Kaminari.paginate_array(@data).page(params[:page]).per(params[:count])
-        render json: { success: true, pagination: set_pagination(:all_sellers), sellers: @all_sellers }
+        render json: {success: true, pagination: set_pagination(:all_sellers), sellers: @all_sellers}
       end
 
 =begin
@@ -103,9 +104,9 @@ module Api
             end
           end
           @all_company_record_on_the_basis_of_roles = Kaminari.paginate_array(@data).page(params[:page]).per(params[:count])
-          render json: { success: true, pagination: set_pagination(:all_company_record_on_the_basis_of_roles), company_record_on_the_basis_of_roles: @all_company_record_on_the_basis_of_roles }
+          render json: {success: true, pagination: set_pagination(:all_company_record_on_the_basis_of_roles), company_record_on_the_basis_of_roles: @all_company_record_on_the_basis_of_roles}
         else
-          render json: { errors: "Not authenticated", response_code: 201 }
+          render json: {errors: "Not authenticated", response_code: 201}
         end
       end
 
@@ -126,16 +127,16 @@ module Api
       def assigned_parcels
         if current_company
           @data = []
-          @parcels = TradingParcel.all.select { |e| e.broker_ids.include? current_company.id.to_s unless e.broker_ids.nil? }
+          @parcels = TradingParcel.all.select {|e| e.broker_ids.include? current_company.id.to_s unless e.broker_ids.nil?}
           @parcels.each do |p|
             @data << {
-              id: p.id,
-              description: p.description
+                id: p.id,
+                description: p.description
             }
           end
-          render json: { success: true, parcels: @data }
+          render json: {success: true, parcels: @data}
         else
-          render json: { errors: "Not authenticated", response_code: 201 }
+          render json: {errors: "Not authenticated", response_code: 201}
         end
       end
 
@@ -158,36 +159,28 @@ module Api
 =end
 
       def send_request
-        if current_company
-          company = Company.where("name LIKE ?", params[:company]).first
-          if company.present?
-            if current_company.is_broker?
-              request = BrokerRequest.where(broker_id: current_company.id,
-                                                   seller_id: company.id,
-                                                   sender_id: current_company.id,
-                                                   receiver_id: company.id).first_or_initialize do |br|
-                br.accepted = false
-              end
-            else
-              request = BrokerRequest.where(broker_id: company.id,
-                                                   seller_id: current_company.id,
-                                                   sender_id: current_company.id,
-                                                   receiver_id: company.id).first_or_initialize do |br|
-                br.accepted = false
-              end
-            end
-            if request.save
-              Message.create_new_broker(request, current_company) if current_company.is_broker?
-              Message.create_new_seller(request, current_company) unless current_company.is_broker?
-              render json: {success: true, message: "Request sent successfully"}
-            else
-              render json: {success: false, message: "Unable to send request."}
-            end
-          else
-            render json: {success: false, message: "There is no company with this name"}
+        company = Company.where("name LIKE ?", params[:company]).first
+        if current_company.is_broker?
+          request = BrokerRequest.where(broker_id: current_company.id,
+                                        seller_id: company.id,
+                                        sender_id: current_company.id,
+                                        receiver_id: company.id).first_or_initialize do |br|
+            br.accepted = false
           end
         else
-          render json: {errors: "Not authenticated", response_code: 201}
+          request = BrokerRequest.where(broker_id: company.id,
+                                        seller_id: current_company.id,
+                                        sender_id: current_company.id,
+                                        receiver_id: company.id).first_or_initialize do |br|
+            br.accepted = false
+          end
+        end
+        if request.save
+          Message.create_new_broker(request, current_company) if current_company.is_broker?
+          Message.create_new_seller(request, current_company) unless current_company.is_broker?
+          render json: {success: true, message: "Request sent successfully"}
+        else
+          render json: {success: false, message: "Unable to send request."}
         end
       end
 
@@ -247,23 +240,23 @@ module Api
           if @parcel.present?
             @demand = Demand.where(description: @parcel.description, block: false).where.not(company_id: current_company.id)
             if @demand.present?
-              all_companies = @demand.map{|d| d.company}
+              all_companies = @demand.map {|d| d.company}
               all_companies.each do |company|
                 @data << {
-                  id: company.id,
-                  name: company.try(:name),
-                  mobile_no: company.try(:get_owner).try(:mobile_no)
+                    id: company.id,
+                    name: company.try(:name),
+                    mobile_no: company.try(:get_owner).try(:mobile_no)
                 }
               end
-              render json: { success: true, companies: @data }
+              render json: {success: true, companies: @data}
             else
-              render json: { success: false, message: "This parcel is not demanded" }
+              render json: {success: false, message: "This parcel is not demanded"}
             end
-          else 
-            render json: { success: false, message: "Parcel for this id does not present." }
+          else
+            render json: {success: false, message: "Parcel for this id does not present."}
           end
         else
-          render json: { errors: "Not authenticated", response_code: 201 }
+          render json: {errors: "Not authenticated", response_code: 201}
         end
       end
 
@@ -287,9 +280,9 @@ module Api
 =end
 
       def accept
-        render json: { success: true,
-                       message: 'Request accepted successfully.',
-                       response_code: 200 } if @broker_request.update_column(:accepted, true)
+        render json: {success: true,
+                      message: 'Request accepted successfully.',
+                      response_code: 200} if @broker_request.update_column(:accepted, true)
       end
 
 =begin
@@ -312,9 +305,9 @@ module Api
 =end
 
       def reject
-        render json: { success: true,
-                       message: 'Request rejected successfully.',
-                       response_code: 200 } if @broker_request.destroy
+        render json: {success: true,
+                      message: 'Request rejected successfully.',
+                      response_code: 200} if @broker_request.destroy
       end
 
 =begin
@@ -337,11 +330,16 @@ module Api
 =end
 
       def remove
-        if @broker_request.destroy
-          CustomerMailer.remove_broker_mail(@broker_request).deliver
-          render json: { success: true,
-                         message: 'You have removed successfully.',
-                         response_code: 200 }
+        if current_company
+          @broker_request = BrokerRequest.find_by(id: params[:request_id], accepted: true)
+          if @broker_request.destroy
+            CustomerMailer.remove_broker_mail(@broker_request).deliver
+            render json: {success: true,
+                          message: 'You have removed successfully.',
+                          response_code: 200}
+          end
+        else
+          render json: {errors: "Not authenticated", response_code: 201}
         end
       end
 
@@ -399,12 +397,12 @@ module Api
             end unless @requests.nil?
           end
           @all_requests = Kaminari.paginate_array(@requests_data).page(params[:page]).per(params[:count])
-          render json: { success: true,
-                         pagination: set_pagination(:all_requests),
-                         response_code: 200,
-                         requests: @requests_data}
+          render json: {success: true,
+                        pagination: set_pagination(:all_requests),
+                        response_code: 200,
+                        requests: @requests_data}
         else
-          render json: { errors: "Not authenticated", response_code: 201 }
+          render json: {errors: "Not authenticated", response_code: 201}
         end
       end
 
@@ -460,12 +458,12 @@ module Api
             end unless @my_brokers.nil?
           end
           @all_clients = Kaminari.paginate_array(@my_clients_data).page(params[:page]).per(params[:count])
-          render json: { success: true,
-                         pagination: set_pagination(:all_clients),
-                         response_code: 200,
-                         myclients: @my_clients_data }
+          render json: {success: true,
+                        pagination: set_pagination(:all_clients),
+                        response_code: 200,
+                        myclients: @my_clients_data}
         else
-          render json: { errors: "Not authenticated", response_code: 201 }
+          render json: {errors: "Not authenticated", response_code: 201}
         end
       end
 
@@ -505,10 +503,10 @@ module Api
 =end
       def dashboard
         if current_company
-            @parcels = TradingParcel.all.select { |e| e.broker_ids.include? current_company.id.to_s unless e.broker_ids.nil? }
-            render status: :ok, template: "api/v1/brokers/dashboard.json.jbuilder"
+          @parcels = TradingParcel.all.select {|e| e.broker_ids.include? current_company.id.to_s unless e.broker_ids.nil?}
+          render status: :ok, template: "api/v1/brokers/dashboard.json.jbuilder"
         else
-            render json: { errors: "Not authenticated", response_code: 201 }
+          render json: {errors: "Not authenticated", response_code: 201}
         end
       end
 
@@ -518,10 +516,31 @@ module Api
         if current_company
           @broker_request = BrokerRequest.find_by(id: params[:request_id], receiver_id: current_company.id)
           if @broker_request.nil?
-            render json: { success: true, message: "#{current_company.name} don't have any request with this id #{params[:request_id]}" }
+            render json: {success: true, message: "#{current_company.name} don't have any request with this id #{params[:request_id]}"}
           end
         else
-          render json: { errors: "Not authenticated", response_code: 201 }
+          render json: {errors: "Not authenticated", response_code: 201}
+        end
+      end
+
+      def check_record_exit?
+        if current_company
+          company = Company.where("name LIKE ?", params[:company]).first
+          if company.present?
+            if current_company.is_broker?
+              request = BrokerRequest.find_by(seller_id: company.id, broker_id: current_company.id)
+            else
+              request = BrokerRequest.find_by(seller_id: current_company.id, broker_id: company.id)
+            end
+            if request.present?
+              render json: {errors: "You both are already connected.", response_code: 201} if request.accepted?
+              render json: {errors: "status is already requested.", response_code: 201} unless request.accepted?
+            end
+          else
+            render json: {success: false, message: "There is no company with this name"}
+          end
+        else
+          render json: {errors: "Not authenticated", response_code: 201}
         end
       end
